Complete renaming.
