+++ magichttp/h1impl/__init__.py
@@ -14,3 +14,5 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
+
+from .impls import H1ClientImpl, H1ServerImpl  # noqa: F401
